Web-Database development/student-management/backend/requirements.txt
flask==2.0.1
flask-sqlalchemy==2.5.1
flask-cors==3.0.10
mysqlclient==2.0.3
python-dotenv==0.19.0 